[FIX] Correct behaviour with precision_rounding = 0 in float utils
`precision_rounding = 0` should mean "as much precision as possible", and not cause
unexpected incorrect calculations. In particular, rounding with a precision of 0 should not
change the input. The same is true if a huge number of digits of precision are requested
(over 323). There is also no situation where it makes sense to say that the literal value
0 is not zero, regardless of how much precision is requested.

This fixes `float_round` and `float_is_zero` for both of these issues.

An alternative would be to raise a `ValueError` if given `precision_rounding` of zero, or a
value of `precision_digits` high enough to have the same effect. I see no advantage in
that unless it is suspected that many users of these functions are already passing in a
value of zero, which hopefully is not the case, since they'll be getting wrong outputs!